Removed Select from client call
If messages were returned that had no From value, the Linq processing in
the original code would throw an exception. Moving the Select to after
the ExecuteAsync fixes this.

It's not ideal becaues now the request on the wire is getting the entire
JSON representation of the message instead of just the fields we care
about, but it fixes the crash.

Resolves #6